Starting Lobby Activity + Moved Files
