+++ app/src/main/res/layout/fragment_lobbyservice_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".LobbyServiceFragment"
+    tools:context=".network.wifi.LobbyServiceFragment"
     tools:listitem="@layout/fragment_lobbyservice" />
